Arrange the search provider and test files.
